Modified register handler

+++ server/src/database/crud/db_crud/database_delete.c
@@ -1,12 +1,12 @@
 #include "../../../../inc/server.h"
 
-void database_delete(const char* from, const char* where) {
+bool database_delete(const char* from, const char* where) {
     sqlite3* db;
     char* error = NULL;
     char* command = NULL;
     int rc = sqlite3_open(exe_path, &db);
 
-    validate_database_operation(rc, db, NULL);
+    if(!validate_database_operation(rc, db, NULL)) return false;
 
     if (where == NULL) {
         asprintf(&command, "DELETE FROM %s", from);
@@ -15,8 +15,11 @@ void database_delete(const char* from, const char* where) {
         asprintf(&command, "DELETE FROM %s WHERE %s", from, where);
     }
     rc = sqlite3_exec(db, command, 0, 0, &error);
-    validate_database_operation(rc, db, error);
+    
+    if(!validate_database_operation(rc, db, error)) return false;
 
     mx_strdel(&command);
     sqlite3_close(db);
+
+    return true;
 }
