+++ server/src/database/migrations/users_groups_migration.c
@@ -5,7 +5,7 @@ void users_groups_migration_up(void) {
     char* error = NULL;
     int rc = sqlite3_open(exe_path, &db);
 
-    validate_database_operation(rc, db, NULL);
+    if(!validate_database_operation(rc, db, NULL)) return;
 
     rc = sqlite3_exec(db,
         "CREATE TABLE IF NOT EXISTS users_groups ("
@@ -17,6 +17,7 @@ void users_groups_migration_up(void) {
         "FOREIGN KEY(group_id) REFERENCES groups(id),"
         "CONSTRAINT UC_users_groups UNIQUE (user_id, group_id))"
         , 0, 0, &error);
+
     validate_database_operation(rc, db, error);
 
     sqlite3_close(db);
@@ -27,7 +28,7 @@ void users_groups_migration_down(void) {
     char* error = NULL;
     int rc = sqlite3_open(exe_path, &db);
 
-    validate_database_operation(rc, db, NULL);
+    if(!validate_database_operation(rc, db, NULL)) return;
 
     rc = sqlite3_exec(db,
         "DROP TABLE IF EXISTS users_groups", 0, 0, &error);
